client page reorg
