+++ app/(app)/system-logs/client-page.tsx
@@ -1,0 +1,1 @@
+export { default as SystemLogsClientPage } from "@/components/system-logs/system-logs-viewer";
